+++ jest.config.js
@@ -1,4 +1,8 @@
 module.exports = {
   verbose: true,
-  collectCoverage: true
+  collectCoverage: true,
+  collectCoverageFrom: [
+    "**/*.{js,jsx,ts,tsx}",
+    "!**/node_modules/**"
+  ]
 }